Создает новый контроллер point.js
